fja-fji created
run.py is seperated into these two
